// packages/twenty-server/src/engine/core-modules/view/enums/view-filter-operand.ts
export enum ViewFilterOperand {
  IS = 'IS',
  IS_NOT_NULL = 'IS_NOT_NULL',
  IS_NOT = 'IS_NOT',
  LESS_THAN_OR_EQUAL = 'LESS_THAN_OR_EQUAL',
  GREATER_THAN_OR_EQUAL = 'GREATER_THAN_OR_EQUAL',
  IS_BEFORE = 'IS_BEFORE',
  IS_AFTER = 'IS_AFTER',
  CONTAINS = 'CONTAINS',
  DOES_NOT_CONTAIN = 'DOES_NOT_CONTAIN',
  IS_EMPTY = 'IS_EMPTY',
  IS_NOT_EMPTY = 'IS_NOT_EMPTY',
  IS_RELATIVE = 'IS_RELATIVE',
  IS_IN_PAST = 'IS_IN_PAST',
  IS_IN_FUTURE = 'IS_IN_FUTURE',
  IS_TODAY = 'IS_TODAY',
  VECTOR_SEARCH = 'VECTOR_SEARCH',
}